Add support for Semaphore 2.0
The new version of Semaphore changed some of the environment variables.

Unfortunately, there's no explicit way to determine if you're using 1.0 or 2.0,
so this PR checks for the original variables first, then the 2.0 variables.

For Branch information, this PR relies on the 'Workflow ID'. This means that
Github status is updated for the same "Branch" across all pipelines in the same
Semaphore 2.0 run.  From their docs, this seems to most closely match how
Semaphore integrates with Github.  The alternative would be to use Pipeline ID.